requirements.txt
# Basics
discord.py
hjson

# Roll plugin
dice